requirements.txt
astroid==1.4.8
colorama==0.3.7
Corpora==1.0
Django==1.10.3
django-bower==5.2.0
django-filter==1.0.1
django-qsstats-magic==0.7.2
django-rest==0.0.1
djangorestframework==3.5.3
docutils==0.12
HTMLParser==0.0.2
isort==4.2.5
lazy-object-proxy==1.2.2
lockfile==0.12.2
Markdown==2.6.7
mccabe==0.5.2
nltk==3.2.1
node==0.9.16
npm==0.1.1
numpy==1.11.2
oauthlib==2.0.1
odict==1.5.2
optional-django==0.1.0
pandas==0.19.1
plumber==1.3.1
pylint==1.6.4
python-daemon==2.1.2
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
requests==2.12.1
requests-oauthlib==0.7.0
selenium==3.0.2
six==1.10.0
tweepy==3.5.0
wrapt==1.10.8
zope.component==4.3.0
zope.deprecation==4.2.0
zope.event==4.2.0
zope.interface==4.3.2
zope.lifecycleevent==4.1.0
